Clean up clippy findings and enforce clippy in CI
Resolve every clippy finding so `cargo clippy --workspace --all-targets
-- -D warnings` is clean, then add it as a required CI step alongside fmt
and test.

- Auto-fixable lints applied via `cargo clippy --fix` across crates,
  examples, and tests: unused_mut, collapsible_if, needless_question_mark,
  bool_assert_comparison, unnecessary/useless conversions, unnecessary_cast,
  explicit_auto_deref, int_plus_one, doc_lazy_continuation.
- new_without_default: derive Default-via-new for RuntimeRegistry and
  DatabaseBuilder.
- approx_constant: test assertions equal to a named constant now use
  std::f64::consts::{PI, LOG10_2}; arbitrary test floats that merely
  resembled pi were retargeted; the STRINGTONUMBER Cosmos-doc example keeps
  3.14 under a scoped allow.
- datetime: replace an infallible i128::try_from in a let-else with i128::from.
- sargable bson_to_raw: keep the explicit match over `.ok()` (crate
  no-silent-discard convention) under a scoped allow.
- slate-store: annotate the PrefixIter lifetime transmutes.

+++ crates/slate-vm/src/lua/convert.rs
@@ -381,5 +381,5 @@ fn is_array_table(table: &mlua::Table) -> Result<bool, mlua::Error> {
             _ => return Ok(false),
         }
     }
-    Ok(count == raw_len as usize)
+    Ok(count == raw_len)
 }
